requirements.txt
disnake
python-dotenv
sqlalchemy
asyncpg
aiohttp
